requirements.txt
torch==2.0.1
torchvision==0.15.2
PyYAML==6.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
Pillow==10.0.0
tqdm==4.65.0
seaborn==0.12.2
matplotlib==3.7.2